Extract PipelineJobItemAPI and impl
